+++ src/lib/data-service.ts
@@ -1,47 +1,39 @@
-import fs from 'fs/promises';
-import path from 'path';
-import { BaseContent } from '@/types/content';
+import { MusicLink, Event, SocialLink, CommunityPost } from '@/types/content';
 
-const DATA_DIR = path.join(process.cwd(), 'src', 'data');
+// Import local JSON data
+import spotifyLinksData from '@/data/spotify-links.json';
+import soundcloudLinksData from '@/data/soundcloud-links.json';
+import youtubeLinksData from '@/data/youtube-links.json';
+import eventsData from '@/data/events.json';
+import socialLinksData from '@/data/social-links.json';
+import communityPostsData from '@/data/community-posts.json';
 
-export async function readData<T extends BaseContent>(filename: string): Promise<T[]> {
-  const filePath = path.join(DATA_DIR, filename);
-  try {
-    const data = await fs.readFile(filePath, 'utf-8');
-    return JSON.parse(data);
-  } catch (error) {
-    console.error(`Error reading ${filename}:`, error);
-    return [];
-  }
+// Spotify Links
+export async function getSpotifyLinks(): Promise<MusicLink[]> {
+  return spotifyLinksData as MusicLink[];
 }
 
-export async function writeData<T extends BaseContent>(filename: string, data: T[]): Promise<boolean> {
-  const filePath = path.join(DATA_DIR, filename);
-  try {
-    await fs.writeFile(filePath, JSON.stringify(data, null, 2));
-    return true;
-  } catch (error) {
-    console.error(`Error writing ${filename}:`, error);
-    return false;
-  }
+// SoundCloud Links
+export async function getSoundCloudLinks(): Promise<MusicLink[]> {
+  return soundcloudLinksData as MusicLink[];
 }
 
-// Helper function to get the appropriate filename for a content type
-export function getFilename(contentType: string): string {
-  switch (contentType) {
-    case 'spotify':
-      return 'spotify-links.json';
-    case 'soundcloud':
-      return 'soundcloud-links.json';
-    case 'youtube':
-      return 'youtube-links.json';
-    case 'events':
-      return 'events.json';
-    case 'social':
-      return 'social-links.json';
-    case 'community':
-      return 'community-posts.json';
-    default:
-      throw new Error(`Unknown content type: ${contentType}`);
-  }
+// YouTube Links
+export async function getYouTubeLinks(): Promise<MusicLink[]> {
+  return youtubeLinksData as MusicLink[];
+}
+
+// Events
+export async function getEvents(): Promise<Event[]> {
+  return eventsData as Event[];
+}
+
+// Social Links
+export async function getSocialLinks(): Promise<SocialLink[]> {
+  return socialLinksData as SocialLink[];
+}
+
+// Community Posts
+export async function getCommunityPosts(): Promise<CommunityPost[]> {
+  return communityPostsData as CommunityPost[];
 }
